+++ packages/explorer-ui/components/layouts/MainLayout/WalletNetworkSection.tsx
@@ -4,7 +4,7 @@
 // import ChainManagement from '@layouts/ChainManagement'
 // import { useTerraWallet } from '@hooks/terra/useTerraWallet'
 
-export default function WalletNetworkSection() {
+export const WalletNetworkSection = () => {
   // const { account } = useActiveWeb3React()
   // const { terraAddress } = useTerraWallet()