+++ packages/explorer-ui/components/misc/StyleAddress.tsx
@@ -1,7 +1,7 @@
 import { ellipsizeString } from '@utils/ellipsizeString'
 import { getAddressesUrl } from '@urls'
 
-export function StyleAddress({ sourceInfo, limiter = 4 }) {
+export const StyleAddress = ({ sourceInfo, limiter = 4 }) => {
   if (sourceInfo.address) {
     return (
       <a
@@ -16,11 +16,12 @@ export function StyleAddress({ sourceInfo, limiter = 4 }) {
           string: sourceInfo.address,
           limiter,
           isZeroX: true,
-        }).slice(0, 4) + ellipsizeString({
-          string: sourceInfo.address,
-          limiter,
-          isZeroX: true,
-        }).slice(6)}
+        }).slice(0, 4) +
+          ellipsizeString({
+            string: sourceInfo.address,
+            limiter,
+            isZeroX: true,
+          }).slice(6)}
       </a>
     )
   } else {