+++ packages/explorer-ui/components/tailwind/Tabs.tsx
@@ -1,6 +1,6 @@
-import Grid from '@components/tailwind/Grid'
+import { Grid } from '@components/tailwind/Grid'
 
-export default function Tabs({ children, numTabs = 2, ...props }) {
+export const Tabs = ({ children, numTabs = 2, ...props }) => {
   return (
     <Grid gap={2} aria-label="Tabs" cols={{ xs: 2, sm: numTabs }} {...props}>
       {children}