+++ packages/synapse-interface/constants/bridgeMap.ts
@@ -1795,18 +1795,32 @@ export const BRIDGE_MAP = {
   },
   '59144': {
     '0x176211869cA2b568f2A7D4EE941E073a821EE1ff': {
+      decimals: 6,
+      symbol: 'USDC',
       origin: ['RFQ.USDC'],
       destination: ['RFQ.USDC'],
-      swappable: [],
-      symbol: 'USDC',
+      swappable: ['0xA219439258ca9da29E9Cc4cE5596924745e12B93'],
+    },
+    '0xA219439258ca9da29E9Cc4cE5596924745e12B93': {
       decimals: 6,
+      symbol: 'USDT',
+      origin: ['RFQ.USDC'],
+      destination: [],
+      swappable: ['0x176211869cA2b568f2A7D4EE941E073a821EE1ff'],
     },
     '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE': {
+      decimals: 18,
+      symbol: 'ETH',
       origin: ['RFQ.ETH'],
       destination: ['RFQ.ETH'],
       swappable: [],
-      symbol: 'ETH',
+    },
+    '0xe5D7C2a44FfDDf6b295A15c148167daaAf5Cf34f': {
       decimals: 18,
+      symbol: 'WETH',
+      origin: ['RFQ.ETH'],
+      destination: [],
+      swappable: [],
     },
   },
   '81457': {
@@ -1820,8 +1834,8 @@ export const BRIDGE_MAP = {
     '0x4300000000000000000000000000000000000003': {
       decimals: 18,
       symbol: 'USDB',
-      origin: ['nUSD', 'RFQ.USDB'],
-      destination: ['nUSD', 'RFQ.USDB'],
+      origin: ['nUSD'],
+      destination: ['nUSD'],
       swappable: ['0x3194B0A295D87fDAA54DF852c248F7a6BAF6c6e0'],
     },
     '0x4300000000000000000000000000000000000004': {
@@ -1864,18 +1878,32 @@ export const BRIDGE_MAP = {
   },
   '534352': {
     '0x06eFdBFf2a14a7c8E15944D1F4A48F9F95F663A4': {
+      decimals: 6,
+      symbol: 'USDC',
       origin: ['RFQ.USDC'],
       destination: ['RFQ.USDC'],
+      swappable: ['0xf55BEC9cafDbE8730f096Aa55dad6D22d44099Df'],
+    },
+    '0x5300000000000000000000000000000000000004': {
+      decimals: 18,
+      symbol: 'WETH',
+      origin: ['RFQ.ETH'],
+      destination: [],
       swappable: [],
-      symbol: 'USDC',
-      decimals: 6,
     },
     '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE': {
+      decimals: 18,
+      symbol: 'ETH',
       origin: ['RFQ.ETH'],
       destination: ['RFQ.ETH'],
       swappable: [],
-      symbol: 'ETH',
-      decimals: 18,
+    },
+    '0xf55BEC9cafDbE8730f096Aa55dad6D22d44099Df': {
+      decimals: 6,
+      symbol: 'USDT',
+      origin: ['RFQ.USDC'],
+      destination: [],
+      swappable: ['0x06eFdBFf2a14a7c8E15944D1F4A48F9F95F663A4'],
     },
   },
   '1313161554': {